outfitter: mark as visited upon enter
in order to track which outfitters the player as visited, the infrastructure
for loading and saving was added.

when the player enters an outfitter for the first time, update
in the game's state that it was visited.

the current implementation keeps the current behavior as is and only sets
up the framework

commit history:
initially, the data structure was implemented as std::set<const Planet *>
as visitedPlanets.
but following tehhowch's suggestion, it was merged into visitedPlanets
and resulted with the conversion of visitedPlanets to std::map<const Planet *, bool>
the logic is both size management and simplified saving flow (in later
commit)

the save to file was reworked (as Hurleveur suggested earlier) so every
outfitter visited will result with the word "outfitter" added after the
name when visited planet option is written to the file.

Signed-off-by: Dagg Stompler <daggs@gmx.com>