missions: add support for cargo in on condition
sometimes a mission requires one to go to a location
an pick up cargo. the current implementation make it possible
by adding the cargo from the start and assuming the player won't
look at it's cargo inventory before arriving to the destination.
this results in a bug when failing the mission prior to the pickup
as the player will have the cargo in his ship without even picking it
up.

this commit add support for cargo pick up as part of the on conditions
set.